requirements.txt
Flask==0.11.1
Pillow==3.4.2
requests==2.9.1
